+++ Sources/StreamChatUI/ChatMessageList/ChatMessageListVCDataSource.swift
@@ -7,8 +7,11 @@ import StreamChat
 
 /// The object that acts as the data source of the message list.
 public protocol ChatMessageListVCDataSource: AnyObject {
-    /// Asks the data source if the first page is currently loaded.
+    /// Asks the data source if the first (newest) page is currently loaded.
     var isFirstPageLoaded: Bool { get }
+    
+    /// Asks the data source if the last (oldest) page is currently loaded.
+    var isLastPageLoaded: Bool { get }
 
     /// Asks the data source to return all the available messages.
     var messages: [ChatMessage] { get set }
@@ -40,6 +43,14 @@ public protocol ChatMessageListVCDataSource: AnyObject {
     ) -> ChatMessageLayoutOptions
 }
 
+/// Default implementations to avoid breaking changes.
+/// Ideally should be implemented by customers who use the `ChatMessageListVC` directly (Advanced).
 public extension ChatMessageListVCDataSource {
-    var isFirstPageLoaded: Bool { true }
+    var isFirstPageLoaded: Bool {
+        true
+    }
+
+    var isLastPageLoaded: Bool {
+        false
+    }
 }